requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.2
cryptography==44.0.1
dj-database-url==3.0.1
Django==5.2
django-allauth==65.4.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.10
google-auth==2.40.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
setuptools==80.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
